separate shared types
