+++ src/types/base/common/shared/types.ts
@@ -1,0 +1,3 @@
+export type NullableString = string | null | undefined;
+export type NullableNumber = number | null | undefined;
+export type NullableBoolean = boolean | null | undefined;
